+++ module/utilities/game-sockets.js
@@ -1,5 +1,6 @@
 import {renderGMIForm} from "../forms/gmi-range-sheet.js";
 import {updateRollDifficultyForm} from "../forms/roll-difficulty-sheet.js";
+import {executeMacroAsGM} from "../macros/macros-scripting.js";
 import {deleteChatMessage, giveAdditionalXP} from "./actor-utilities.js";
 import {resetDifficulty} from "./roll-engine/roll-engine-main.js";
 
@@ -10,5 +11,6 @@ export function gameSockets() {
     if (data.operation === "renderGMIForm") renderGMIForm();
     if (data.operation === "updateRollDifficultyForm") updateRollDifficultyForm();
     if (data.operation === "resetDifficulty") resetDifficulty();
+    if (data.operation === "executeMacroAsGM") executeMacroAsGM(data.macroUuid, data.rollData);
   });
 }